Added item purchase and order mark as paid

+++ AsgardMarketplace.Api/Controllers/ArgardController.cs
@@ -24,24 +24,59 @@ namespace AsgardMarketplace.Api.Controllers
         #region Items
 
         //Gets Item by id
-        [HttpGet("items/{id}")]
-        public async Task<ActionResult<Item>> Get(string id)
+        [HttpGet("items/{Id}")]
+        public async Task<ActionResult<Item>> Get(string Id)
         {
-            id = id.ToLower().Trim();
-            Item item = await repo.GetItem(id);
+            int itemId;
+            try {
+                itemId = int.Parse(Id.Trim());
+            }
+            catch (Exception ex) {
+                return BadRequest(ex.Message);
+            }
+            Item item = await repo.GetItem(itemId);
 
-            if (item.State != "NOT_FOUND")
+            if (item.State == "NOT_FOUND")
             {
-                return NotFound("Item for id:" + id + "not found");
+                return NotFound("Item for id: " + Id + " not found");
             }
-            else
+
+            return Ok(item);
+        }
+
+        //Purchase item
+        [HttpPost("items/purchase/")]
+        public async Task<ActionResult<Item>> Post([FromQuery]string buyerId, [FromQuery]string sellerId, [FromQuery]string itemIdString, [FromQuery]string quantityString)
+        {
+            int itemId;
+            try {
+                itemId = int.Parse(itemIdString.Trim());
+            }
+            catch (Exception ex) {
+                return BadRequest(ex.Message);
+            }
+            int quantity;
+            try {
+                quantity = int.Parse(quantityString.Trim());
+            }
+            catch (Exception ex) {
+                return BadRequest(ex.Message);
+            }
+            Item item = await repo.GetItem(itemId);
+
+            if (item.State == "NOT_FOUND")
             {
-                return Ok(item);
+                return NotFound("Item for id: " + itemIdString + " not found");
             }
+            if (item.Quantity < quantity)
+                return Ok("Not enough quantity");
+            OrderService orderService = new OrderService(repo);
+            var marked = await orderService.CreateOrder(buyerId, sellerId, item, quantity);
+            return Ok(marked);
         }
 
         //Create an item
-        [HttpPost("items/create-item")]
+        [HttpPost("items/createItem")]
         public ActionResult<Item> Post(Item item)
         {
             var itemService = new ItemService(repo);
@@ -59,36 +94,40 @@ namespace AsgardMarketplace.Api.Controllers
         }
 
         //Deletes item
-        [HttpGet("items/delete/{itemId}")]
-        public async Task<ActionResult<Item>> DeleteItem(string itemId)
+        [HttpGet("items/delete/{Id}")]
+        public async Task<ActionResult<Item>> DeleteItem(string Id)
         {
-            Item item = await repo.GetItem(itemId.ToString());
-            repo.deleteItem(item);
+            int itemId;
+            try {
+                itemId = int.Parse(Id.Trim());
+            }
+            catch (Exception ex) {
+                return Ok("Invalid Id");
+            }
+            Item item = await repo.GetItem(itemId);
+            if (item.State != "NOT_FOUND") {
+                repo.deleteItem(item);
+                return Ok("Item with id: " + itemId + " was successfully deleted");
+            }
 
-            return NoContent();
+            return NotFound("Item with id: " + itemId + " was not found");
         }
 
         #endregion
         #region Orders
-
+        //get?id=00&offset=aa
         //Gets all orders for user
-        [HttpGet("orders/{userId}")]
-        public async Task<ActionResult<Order>> GetOrders(string userId, int ofs, int lmt)
+        [HttpGet("orders/")]
+        public async Task<ActionResult<Order>> GetOrders([FromQuery]string Id, [FromQuery]int offset, [FromQuery]int limit)
         {
-
-            List<Order> orders = await repo.GetAllOrders(
-                ofs, //Offset
-                lmt //Limit
-                );
-
-            List<Order> userOrders = new List<Order>();
-            foreach (Order order in orders)
-            {
-                if (order.Buyer == userId) //We only want to return orders for this user
-                {
-                    userOrders.Add(order);
-                }
+            int userId;
+            try {
+                userId = int.Parse(Id.Trim());
             }
+            catch (Exception ex) {
+                return Ok("Invalid user Id");
+            }
+            List<Order> userOrders = await repo.GetAllOrders(offset, limit, userId);
 
             if (userOrders.Count() == 0)
             {
@@ -99,12 +138,20 @@ namespace AsgardMarketplace.Api.Controllers
         }
 
         //Marks order as delivered
-        [HttpPost("markAsDelivered/{orderId}")]
-        public async Task<ActionResult<Order>> MarkOrderAsDelivered(string user_id_1 /*Seller userId*/, string orderId)
+        // Id - Seller user ID
+        [HttpPost("markAsDelivered/")]
+        public async Task<ActionResult<Order>> MarkOrderAsDelivered([FromQuery]string Id, [FromQuery]string orderId)
         {
+            int userId;
+            try {
+                userId = int.Parse(Id.Trim());
+            }
+            catch (Exception ex) {
+                return Ok("Invalid user Id");
+            }
             OrderService orderService = new OrderService(repo);
-
-            var marked = await orderService.MarkOrderAsDelivered(user_id_1, Guid.Parse(orderId));
+            
+            var marked = await orderService.MarkOrderAsDelivered(userId.ToString(), Guid.Parse(orderId));
 
             // If marking the order did not failed we return Ok
             if (marked != false)
@@ -112,6 +159,16 @@ namespace AsgardMarketplace.Api.Controllers
             else
                 return Ok();
 
+        }
+
+        [HttpPost("markAsPaid/")]
+        public async Task<ActionResult<Order>> markAsPaid([FromQuery]string Id, [FromQuery]string paymentId, [FromQuery]string paymentUrl)
+        {
+            OrderService orderService = new OrderService(repo);
+            var result = await orderService.MarkOrderAsPaid(Guid.Parse(Id.Trim()), paymentId, paymentUrl);
+            if (!result)
+                return Ok("An error occured while marking order: " + Id + " as Paid");
+            return Ok("Order with id: " + Id + " marked as paid");
         }
 
         #endregion
